Already turned it in but then I remembered about decimals and changed the input type to floats

+++ ps0.py
@@ -1,9 +1,9 @@
 import math
 # 1. Ask the user to enter a number "x"
-x = int(input('Enter number x:'))
+x = float(input('Enter number x:'))
 
 # 2. Ask the user to enter a number "y"
-y = int(input('Enter number y:'))
+y = float(input('Enter number y:'))
 
 # 3. Prints out the number "x" raised to the power "y"
 print('x^y =',x**y)
